Auto-format (#473)

+++ src/main/java/com/paneedah/weaponlib/particle/vehicle/TurbulentSmokeParticle.java
@@ -12,90 +12,88 @@ import java.util.Random;
 
 public class TurbulentSmokeParticle extends ParticleCloud {
 
-	
-	public TurbulentSmokeParticle(World worldIn, double xCoordIn, double yCoordIn, double zCoordIn, double p_i1221_8_,
-			double p_i1221_10_, double p_i1221_12_) {
-		super(worldIn, xCoordIn, yCoordIn, zCoordIn, p_i1221_8_, p_i1221_10_, p_i1221_12_);
-		setParticleTexture(ClientEventHandler.carParticles);
-		particleScale = 0.5f;
-		
-	}
-	
-	@Override
-	public boolean shouldDisableDepth() {
-		return true;
-	}
-	
-	public TurbulentSmokeParticle(World worldIn, double xCoordIn, double yCoordIn, double zCoordIn, double p_i1221_8_,
-			double p_i1221_10_, double p_i1221_12_, int indexY) {
-		super(worldIn, xCoordIn, yCoordIn, zCoordIn, p_i1221_8_, p_i1221_10_, p_i1221_12_);
-		setParticleTexture(ClientEventHandler.carParticles);
-		particleScale = 0.5f;
-		
-		this.particleTextureIndexY = indexY;
-		
-	}
-	
-	
-	@Override
-	public int getFXLayer() {
-		return 1;
-	}
-	
-	@Override
-	public void renderParticle(BufferBuilder buffer, Entity entityIn, float partialTicks, float rotationX,
-			float rotationZ, float rotationYZ, float rotationXY, float rotationXZ) {
-		
-		
-		//rotationX = 0f;
-		//rotationXZ = 30f;
-			
-			Random urandom = new Random(333);
-			
-			
-			 this.particleRed = this.particleGreen = this.particleBlue = urandom.nextFloat() * 0.5F + 0.4F;
-		       
-		        int j = this.getBrightnessForRender(partialTicks);
-		        int k = j >> 16 & 65535;
-		        int l = j & 65535;
-		        
-				float scale = this.particleScale;
-				float pX = (float)(this.prevPosX + (this.posX - this.prevPosX) * (double)partialTicks - interpPosX);
-		        float pY = (float)(this.prevPosY + (this.posY - this.prevPosY) * (double)partialTicks - interpPosY);
-		        float pZ = (float)(this.prevPosZ + (this.posZ - this.prevPosZ) * (double)partialTicks - interpPosZ);
-		        
-		        double minX = particleTexture.getMinU() + ((particleTexture.getMaxU()-particleTexture.getMinU()) * (particleTextureIndexX/8f));
-		        double minY = particleTexture.getMinV() + ((particleTexture.getMaxV()-particleTexture.getMinV()) * (particleTextureIndexY/8f));
-		        double mU = (particleTexture.getMaxU()-particleTexture.getMinU())/8;
-		        double mV = (particleTexture.getMaxV()-particleTexture.getMinV())/8;
-		        
-		        buffer.pos((double)(pX - rotationX * scale - rotationXY * scale), (double)(pY - rotationZ * scale), (double)(pZ - rotationYZ * scale - rotationXZ * scale)).tex(minX+mU, minY+mV).color(this.particleRed, this.particleGreen, this.particleBlue, this.particleAlpha).lightmap(k, l).endVertex();
-				buffer.pos((double)(pX - rotationX * scale + rotationXY * scale), (double)(pY + rotationZ * scale), (double)(pZ - rotationYZ * scale + rotationXZ * scale)).tex(minX+mU, minY).color(this.particleRed, this.particleGreen, this.particleBlue, this.particleAlpha).lightmap(k, l).endVertex();
-				buffer.pos((double)(pX + rotationX * scale + rotationXY * scale), (double)(pY + rotationZ * scale), (double)(pZ + rotationYZ * scale + rotationXZ * scale)).tex(minX, minY).color(this.particleRed, this.particleGreen, this.particleBlue, this.particleAlpha).lightmap(k, l).endVertex();
-				buffer.pos((double)(pX + rotationX * scale - rotationXY * scale), (double)(pY - rotationZ * scale), (double)(pZ + rotationYZ * scale - rotationXZ * scale)).tex(minX, minY+mV).color(this.particleRed, this.particleGreen, this.particleBlue, this.particleAlpha).lightmap(k, l).endVertex();
-			
-		
-	}
-	
-	
-	@Override
-	public void onUpdate() {
-		this.prevPosX = this.posX;
+
+    public TurbulentSmokeParticle(World worldIn, double xCoordIn, double yCoordIn, double zCoordIn, double p_i1221_8_,
+                                  double p_i1221_10_, double p_i1221_12_) {
+        super(worldIn, xCoordIn, yCoordIn, zCoordIn, p_i1221_8_, p_i1221_10_, p_i1221_12_);
+        setParticleTexture(ClientEventHandler.carParticles);
+        particleScale = 0.5f;
+
+    }
+
+    @Override
+    public boolean shouldDisableDepth() {
+        return true;
+    }
+
+    public TurbulentSmokeParticle(World worldIn, double xCoordIn, double yCoordIn, double zCoordIn, double p_i1221_8_,
+                                  double p_i1221_10_, double p_i1221_12_, int indexY) {
+        super(worldIn, xCoordIn, yCoordIn, zCoordIn, p_i1221_8_, p_i1221_10_, p_i1221_12_);
+        setParticleTexture(ClientEventHandler.carParticles);
+        particleScale = 0.5f;
+
+        this.particleTextureIndexY = indexY;
+
+    }
+
+
+    @Override
+    public int getFXLayer() {
+        return 1;
+    }
+
+    @Override
+    public void renderParticle(BufferBuilder buffer, Entity entityIn, float partialTicks, float rotationX,
+                               float rotationZ, float rotationYZ, float rotationXY, float rotationXZ) {
+
+
+        //rotationX = 0f;
+        //rotationXZ = 30f;
+
+        Random urandom = new Random(333);
+
+
+        this.particleRed = this.particleGreen = this.particleBlue = urandom.nextFloat() * 0.5F + 0.4F;
+
+        int j = this.getBrightnessForRender(partialTicks);
+        int k = j >> 16 & 65535;
+        int l = j & 65535;
+
+        float scale = this.particleScale;
+        float pX = (float) (this.prevPosX + (this.posX - this.prevPosX) * (double) partialTicks - interpPosX);
+        float pY = (float) (this.prevPosY + (this.posY - this.prevPosY) * (double) partialTicks - interpPosY);
+        float pZ = (float) (this.prevPosZ + (this.posZ - this.prevPosZ) * (double) partialTicks - interpPosZ);
+
+        double minX = particleTexture.getMinU() + ((particleTexture.getMaxU() - particleTexture.getMinU()) * (particleTextureIndexX / 8f));
+        double minY = particleTexture.getMinV() + ((particleTexture.getMaxV() - particleTexture.getMinV()) * (particleTextureIndexY / 8f));
+        double mU = (particleTexture.getMaxU() - particleTexture.getMinU()) / 8;
+        double mV = (particleTexture.getMaxV() - particleTexture.getMinV()) / 8;
+
+        buffer.pos(pX - rotationX * scale - rotationXY * scale, pY - rotationZ * scale, pZ - rotationYZ * scale - rotationXZ * scale).tex(minX + mU, minY + mV).color(this.particleRed, this.particleGreen, this.particleBlue, this.particleAlpha).lightmap(k, l).endVertex();
+        buffer.pos(pX - rotationX * scale + rotationXY * scale, pY + rotationZ * scale, pZ - rotationYZ * scale + rotationXZ * scale).tex(minX + mU, minY).color(this.particleRed, this.particleGreen, this.particleBlue, this.particleAlpha).lightmap(k, l).endVertex();
+        buffer.pos(pX + rotationX * scale + rotationXY * scale, pY + rotationZ * scale, pZ + rotationYZ * scale + rotationXZ * scale).tex(minX, minY).color(this.particleRed, this.particleGreen, this.particleBlue, this.particleAlpha).lightmap(k, l).endVertex();
+        buffer.pos(pX + rotationX * scale - rotationXY * scale, pY - rotationZ * scale, pZ + rotationYZ * scale - rotationXZ * scale).tex(minX, minY + mV).color(this.particleRed, this.particleGreen, this.particleBlue, this.particleAlpha).lightmap(k, l).endVertex();
+
+
+    }
+
+
+    @Override
+    public void onUpdate() {
+        this.prevPosX = this.posX;
         this.prevPosY = this.posY;
         this.prevPosZ = this.posZ;
 
-        
-        
-        if (this.particleAge++ >= this.particleMaxAge)
-        {
-        	 this.setExpired();
+
+        if (this.particleAge++ >= this.particleMaxAge) {
+            this.setExpired();
         }
-        
+
         this.particleTextureIndexX = this.particleAge * 7 / this.particleMaxAge;
 
-        
+
         Random rand = new Random();
-       
+
         this.move(this.motionX, this.motionY, this.motionZ);
         
        
@@ -136,38 +134,30 @@ public class TurbulentSmokeParticle extends ParticleCloud {
         } catch(Exception e) {
         	//e.printStackTrace();
         };*/
-        
-        
-        
-       
-        
+
+
         this.motionX *= 0.9599999785423279D;
         this.motionY *= 0.9499999785423279D;
         this.motionZ *= 0.9599999785423279D;
-        
-     
-     
+
+
         EntityPlayer entityplayer = this.world.getClosestPlayer(this.posX, this.posY, this.posZ, 2.0D, false);
 
-        if (entityplayer != null)
-        {
+        if (entityplayer != null) {
             AxisAlignedBB axisalignedbb = entityplayer.getEntityBoundingBox();
 
-            if (this.posY > axisalignedbb.minY)
-            {
+            if (this.posY > axisalignedbb.minY) {
                 this.posY += (axisalignedbb.minY - this.posY) * 0.2D;
                 this.motionY += (entityplayer.motionY - this.motionY) * 0.2D;
                 this.setPosition(this.posX, this.posY, this.posZ);
             }
         }
 
-        
-        
-        if (this.onGround)
-        {
-        	this.setExpired();
+
+        if (this.onGround) {
+            this.setExpired();
             this.motionX *= 0.699999988079071D;
             this.motionZ *= 0.699999988079071D;
         }
-	}
+    }
 }
